backend/requirement.txt
opencv-python==4.10.0.84
numpy==1.24.4
pillow==10.2.0
torch>=2.2.0
torchvision>=0.17.0,<0.18.0
facenet-pytorch==2.6.0
supabase==2.8.0
python-dotenv==1.0.0

fastapi==0.104.1
uvicorn==0.24.0

python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
python-dotenv==1.0.0